Add dividend as cash properly
"self.datadiv" is a list of dividends for each tick. Adding the whole
list of dividends on each processing step is erroneous and also causes
various other issues.

Error in backtrader/analyzers/drawdown.py", line 93, in notify_fund:
ValueError: The truth value of an array with more than one element is
ambiguous. Use a.any() or a.all()